Fix commit confirmation logic
`vim.eval('getchar()')` returns a string representation of the character code,
not an integer. The previous `isinstance(..., int)` check would always fail,
making it impossible to confirm a commit.

This change correctly casts the string to an integer before converting it to a
character. It also removes an unused variable and a redundant check,
simplifying the confirmation logic.

Signed-off-by: Simo Sorce <simo@redhat.com>